Widen int multiplications before conversion (benign sites)
CodeQL flags int*int multiplications whose result is converted to a
64-bit type (size_t / unsigned long / address arithmetic), because the
product is computed in 32 bits and can overflow before it is widened
(alerts 59, 60, 61, 62, 66, 67, 97, 99).

Triage of these sites shows none can realistically overflow: the
communicator-allgather emulation multiplies a communicator size by a
constant count of 2; the debugger DLLs multiply a pointer-array index
(bounded by the debuggee's communicator table) by the pointer size;
the affinity extension multiplies node-local socket and core counts;
and the bipartite graph squares a vertex count bounded by NIC
interface counts.

Widen them anyway with an explicit cast so the arithmetic is correct
by construction and the alerts are resolved.  The tmpbuf index
arithmetic next to the widened allgather-emulation allocations is
widened as well, so the index math matches the allocation math.

+++ ompi/communicator/comm.c
@@ -2393,18 +2393,18 @@ static int ompi_comm_allgather_emulate_intra( void *inbuf, int incount,
 
     /* Step 1: the gather-step */
     if ( 0 == rank ) {
-        tmpbuf = (int *) malloc (rsize*outcount*sizeof(int));
+        tmpbuf = (int *) malloc ((size_t) rsize * outcount * sizeof(int));
         if ( NULL == tmpbuf ) {
             return (OMPI_ERR_OUT_OF_RESOURCE);
         }
-        req = (MPI_Request *)malloc (rsize*outcount*sizeof(MPI_Request));
+        req = (MPI_Request *)malloc ((size_t) rsize * outcount * sizeof(MPI_Request));
         if ( NULL == req ) {
             free ( tmpbuf );
             return (OMPI_ERR_OUT_OF_RESOURCE);
         }
 
         for ( i=0; i<rsize; i++) {
-            rc = MCA_PML_CALL(irecv( &tmpbuf[outcount*i], outcount, outtype, i,
+            rc = MCA_PML_CALL(irecv( &tmpbuf[(size_t) outcount * i], outcount, outtype, i,
                                      OMPI_COMM_ALLGATHER_TAG, comm, &req[i] ));
             if ( OMPI_SUCCESS != rc ) {
                 goto exit;
@@ -2430,7 +2430,7 @@ static int ompi_comm_allgather_emulate_intra( void *inbuf, int incount,
     }
 
     /* Step 2: the inter-bcast step */
-    rc = MCA_PML_CALL(irecv (outbuf, size*outcount, outtype, 0,
+    rc = MCA_PML_CALL(irecv (outbuf, (size_t) size * outcount, outtype, 0,
                              OMPI_COMM_ALLGATHER_TAG, comm, &sendreq));
     if ( OMPI_SUCCESS != rc ) {
         goto exit;
@@ -2438,7 +2438,7 @@ static int ompi_comm_allgather_emulate_intra( void *inbuf, int incount,
 
     if ( 0 == rank ) {
         for ( i=0; i < rsize; i++ ){
-            rc = MCA_PML_CALL(send (tmpbuf, rsize*outcount, outtype, i,
+            rc = MCA_PML_CALL(send (tmpbuf, (size_t) rsize * outcount, outtype, i,
                                     OMPI_COMM_ALLGATHER_TAG,
                                     MCA_PML_BASE_SEND_STANDARD, comm));
             if ( OMPI_SUCCESS != rc ) {
